finished porting bare-bones sim2d
