+++ iarc_sim_2d/launch/basic.launch
@@ -6,8 +6,6 @@
 	<include file="$(find iarc_sim_engine)/launch/basic.launch"/>
 
 	<node name="iarc_simulator" pkg="iarc_sim_2d" cwd="node" type="sim.py" output="screen">
-        <param name="num_targets" value="$(arg num_targets)"/>
-        <param name="num_obstacles" value="$(arg num_obstacles)"/>
     </node>
 	
 	<node name="cmd_vel_fwd" pkg="topic_tools" type="relay" args="/cmd_vel /drone/cmd_vel" />
